backend/requirements.txt
Django == 4.2.15
djangorestframework == 3.15.2
markdown == 3.5.1
django-filter == 23.3
mysqlclient == 2.2.0
python-dotenv == 1.0.0
whitenoise == 6.6.0
gunicorn == 22.0.0
requests == 2.32.0
apscheduler == 3.10.4
django-cors-headers == 4.3.1
aiohttp == 3.10.2
adrf == 0.1.3
boto3 == 1.34.84
django-cors-headers==4.3.1
django-extensions == 3.2.3
PyJWT==2.10.1
cryptography==45.0.4